+++ CHANGELOG.md
@@ -4,6 +4,7 @@ UpShot Changelog
 master
 ------
 * Allow auto-scaling of retina screenshots down to 72dpi. Fix #44.
+* Add automated update service (Sparkle!). Fix #14.
 
 v1.1 (4/26/2013)
 ----------------